Changed data type for Native Token quantity from uint to ulong to reflect onchain data type

+++ CardanoSharp.Wallet/TransactionBuilding/NativeAssetBuilder.cs
@@ -12,7 +12,7 @@ namespace CardanoSharp.Wallet.TransactionBuilding
             _model = new NativeAsset();
         }
 
-        public NativeAssetBuilder WithToken(Dictionary<byte[], uint> token)
+        public NativeAssetBuilder WithToken(Dictionary<byte[], ulong> token)
         {
             _model.Token = token;
             return this;
